Add Giurisprudenza retrieving and parsing
* Generify Retriever and Parser

So far it looks like we can reuse the same retriever and the same parser
for bot DST and Giurispridenza.

Going with the flow 👍🏻
